+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ### Fixed:
 - Fixed dispenser gets changed to copper cauldron
 ### Changed:
-- 
+- Tweaked bamboo pump water check